single prec

+++ src/complex_langevin/config.py
@@ -2,7 +2,7 @@
 import torch
 import os
 
-CL_PRECISION = os.getenv("CL_PRECISION", "double").lower()
+CL_PRECISION = os.getenv("CL_PRECISION", "single").lower()
 
 if CL_PRECISION == "double":
     CL_REAL = torch.float64
